// src/app/modules/applications/applications.interface.ts
export interface IApplication {
  jobId: string
  jobTitle: string
  name: string
  email: string
  phone: string
  salaryExpectation: string
  githubUrl: string
  linkedinUrl: string 
  portfolioUrl: string
  resumeLink: string
  location: string
  yearsOfExperience: string
  problemSolvingExperience: string
  whyHireYou: string
  // Additional fields as needed
}